Unified db schema for the app to simplify usings and enabling codegen follow and reference common db schema better
