+++ agent/policies/handlers.py
@@ -58,7 +58,6 @@ interface Output {
 <handler>
 import { db } from "../db";
 import type { Greeting } from "../db/schema/application";
-import { messagesTable } from "../db/schema/common";
 
 export const handle = (options: Options): Output => {
     // Store user message
